refactor realpath handling
