+++ src/resolve-dependency.js
@@ -1,27 +1,4 @@
-const { basename, dirname, isAbsolute, resolve, sep } = require('path');
-
-function inPath (path, parent) {
-  return path.startsWith(parent) && path[parent.length] === sep;
-}
-
-function realpath (path, parent, job, seen = new Set()) {
-  if (seen.has(path)) throw new Error('Recursive symlink detected resolving ' + path);
-  seen.add(path);
-  const symlink = job.readlink(path);
-  // emit direct symlink paths only
-  if (symlink) {
-    const parentPath = dirname(path);
-    const resolved = resolve(parentPath, symlink);
-    const realParent = realpath(parentPath, parent, job);
-    if (inPath(path, realParent))
-      job.emitFile(path, 'resolve', parent);
-    return realpath(resolved, parent, job, seen);
-  }
-  // keep backtracking for realpath, emitting folder symlinks within base
-  if (!inPath(path, job.base))
-    return path;
-  return realpath(dirname(path), parent, job, seen) + sep + basename(path);
-}
+const { isAbsolute, resolve, sep } = require('path');
 
 // node resolver
 // custom implementation to emit only needed package.json files for resolver
@@ -33,7 +10,7 @@ module.exports = function resolveDependency (specifier, parent, job) {
   else
     resolved = resolvePackage(specifier, parent, job);
   if (resolved.startsWith('node:')) return resolved;
-  return realpath(resolved, parent, job);
+  return job.realpath(resolved, parent);
 };
 
 function resolvePath (path, parent, job) {
@@ -52,7 +29,7 @@ function resolveFile (path, job) {
 
 function resolveDir (path, parent, job) {
   if (!job.isDir(path)) return;
-  const realPjsonPath = realpath(path + '/package.json', parent, job);
+  const realPjsonPath = job.realpath(path + '/package.json', parent);
   const pjsonSource = job.readFile(realPjsonPath);
   if (pjsonSource) {
     try {
